♻️(ingestion-presentation) isolate sources views
