+++ src/web/presentation/ingestion/views.py
@@ -58,7 +58,6 @@ from presentation.ingestion.serializers import (
     ListOffersResponseSerializer,
     NoValidRowsErrorSerializer,
     OffersInputSerializer,
-    SourceSerializer,
     TokenErrorSerializer,
     UpsertOffersRequestSerializer,
 )
@@ -433,22 +432,6 @@ class OffersListView(APIView):
             serializer = GenericErrorSerializer({"error": "Unexpected error"})
             return Response(
                 serializer.data, status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-
-
-@extend_schema(exclude=True)
-class SourcesListView(APIView):
-    authentication_classes = [ApiKeyAuthentication]
-
-    def get(self, request):
-        try:
-            container = create_ingestion_container()
-            sources = container.list_sources_usecase().execute()
-            return Response(SourceSerializer(sources, many=True).data)
-        except Exception:
-            return Response(
-                {"error": "Unexpected error"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
